--- core/src/main/java/org/elasticsearch/index/query/morelikethis/Item.java
@@ -1,339 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.elasticsearch.index.query.morelikethis;
-
-import org.elasticsearch.ElasticsearchParseException;
-import org.elasticsearch.ExceptionsHelper;
-import org.elasticsearch.action.termvectors.TermVectorsRequest;
-import org.elasticsearch.common.Nullable;
-import org.elasticsearch.common.ParseField;
-import org.elasticsearch.common.ParseFieldMatcher;
-import org.elasticsearch.common.bytes.BytesReference;
-import org.elasticsearch.common.lucene.uid.Versions;
-import org.elasticsearch.common.xcontent.*;
-import org.elasticsearch.index.VersionType;
-
-import java.io.IOException;
-import java.util.*;
-
-import static org.elasticsearch.common.xcontent.XContentFactory.jsonBuilder;
-
-/**
- * A single item to be used for a {@link MoreLikeThisQueryBuilder}.
- */
-public final class Item implements ToXContent {
-    public static final Item[] EMPTY_ARRAY = new Item[0];
-
-    public interface Field {
-        ParseField INDEX = new ParseField("_index");
-        ParseField TYPE = new ParseField("_type");
-        ParseField ID = new ParseField("_id");
-        ParseField DOC = new ParseField("doc");
-        ParseField FIELDS = new ParseField("fields");
-        ParseField PER_FIELD_ANALYZER = new ParseField("per_field_analyzer");
-        ParseField ROUTING = new ParseField("_routing");
-        ParseField VERSION = new ParseField("_version");
-        ParseField VERSION_TYPE = new ParseField("_version_type");
-    }
-
-    private String index;
-    private String type;
-    private String id;
-    private BytesReference doc;
-    private String[] fields;
-    private Map<String, String> perFieldAnalyzer;
-    private String routing;
-    private long version = Versions.MATCH_ANY;
-    private VersionType versionType = VersionType.INTERNAL;
-
-    public Item() {
-
-    }
-
-    /**
-     * Constructor for a given item / document request
-     *
-     * @param index the index where the document is located
-     * @param type the type of the document
-     * @param id and its id
-     */
-    public Item(String index, @Nullable String type, String id) {
-        this.index = index;
-        this.type = type;
-        this.id = id;
-    }
-
-    /**
-     * Constructor for an artificial document request, that is not present in the index.
-     *
-     * @param index the index to be used for parsing the doc
-     * @param type the type to be used for parsing the doc
-     * @param doc the document specification
-     */
-    public Item(String index, String type, XContentBuilder doc) {
-        this.index = index;
-        this.type = type;
-        this.doc(doc);
-    }
-
-    public String index() {
-        return index;
-    }
-
-    public Item index(String index) {
-        this.index = index;
-        return this;
-    }
-
-    public String type() {
-        return type;
-    }
-
-    public Item type(String type) {
-        this.type = type;
-        return this;
-    }
-
-    public String id() {
-        return id;
-    }
-
-    public Item id(String id) {
-        this.id = id;
-        return this;
-    }
-
-    public BytesReference doc() {
-        return doc;
-    }
-
-    /**
-     * Sets to a given artificial document, that is a document that is not present in the index.
-     */
-    public Item doc(BytesReference doc) {
-        this.doc = doc;
-        return this;
-    }
-
-    /**
-     * Sets to a given artificial document, that is a document that is not present in the index.
-     */
-    public Item doc(XContentBuilder doc) {
-        return this.doc(doc.bytes());
-    }
-
-    public String[] fields() {
-        return fields;
-    }
-
-    public Item fields(String... fields) {
-        this.fields = fields;
-        return this;
-    }
-
-    public Map<String, String> perFieldAnalyzer() {
-        return perFieldAnalyzer;
-    }
-
-    /**
-     * Sets the analyzer(s) to use at any given field.
-     */
-    public Item perFieldAnalyzer(Map<String, String> perFieldAnalyzer) {
-        this.perFieldAnalyzer = perFieldAnalyzer;
-        return this;
-    }
-
-    public String routing() {
-        return routing;
-    }
-
-    public Item routing(String routing) {
-        this.routing = routing;
-        return this;
-    }
-
-    public long version() {
-        return version;
-    }
-
-    public Item version(long version) {
-        this.version = version;
-        return this;
-    }
-
-    public VersionType versionType() {
-        return versionType;
-    }
-
-    public Item versionType(VersionType versionType) {
-        this.versionType = versionType;
-        return this;
-    }
-
-    /**
-     * Convert this to a {@link TermVectorsRequest} for fetching the terms of the document.
-     */
-    public TermVectorsRequest toTermVectorsRequest() {
-        TermVectorsRequest termVectorsRequest = new TermVectorsRequest(index, type, id)
-                .selectedFields(fields)
-                .routing(routing)
-                .version(version)
-                .versionType(versionType)
-                .perFieldAnalyzer(perFieldAnalyzer)
-                .positions(false)  // ensures these following parameters are never set
-                .offsets(false)
-                .payloads(false)
-                .fieldStatistics(false)
-                .termStatistics(false)
-                .dfs(false);
-        // for artificial docs to make sure that the id has changed in the item too
-        if (doc != null) {
-            termVectorsRequest.doc(doc, true);
-            this.id(termVectorsRequest.id());
-        }
-        return termVectorsRequest;
-    }
-
-    /**
-     * Parses and returns the given item.
-     */
-    public static Item parse(XContentParser parser, Item item) throws IOException {
-        XContentParser.Token token;
-        String currentFieldName = null;
-        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
-            if (token == XContentParser.Token.FIELD_NAME) {
-                currentFieldName = parser.currentName();
-            } else if (currentFieldName != null) {
-                if (ParseFieldMatcher.STRICT.match(currentFieldName, Field.INDEX)) {
-                    item.index = parser.text();
-                } else if (ParseFieldMatcher.STRICT.match(currentFieldName, Field.TYPE)) {
-                    item.type = parser.text();
-                } else if (ParseFieldMatcher.STRICT.match(currentFieldName, Field.ID)) {
-                    item.id = parser.text();
-                } else if (ParseFieldMatcher.STRICT.match(currentFieldName, Field.DOC)) {
-                    item.doc(jsonBuilder().copyCurrentStructure(parser));
-                } else if (ParseFieldMatcher.STRICT.match(currentFieldName, Field.FIELDS)) {
-                    if (token == XContentParser.Token.START_ARRAY) {
-                        List<String> fields = new ArrayList<>();
-                        while (parser.nextToken() != XContentParser.Token.END_ARRAY) {
-                            fields.add(parser.text());
-                        }
-                        item.fields(fields.toArray(new String[fields.size()]));
-                    } else {
-                        throw new ElasticsearchParseException(
-                                "failed to parse More Like This item. field [fields] must be an array");
-                    }
-                } else if (ParseFieldMatcher.STRICT.match(currentFieldName, Field.PER_FIELD_ANALYZER)) {
-                    item.perFieldAnalyzer(TermVectorsRequest.readPerFieldAnalyzer(parser.map()));
-                } else if ("_routing".equals(currentFieldName) || "routing".equals(currentFieldName)) {
-                    item.routing = parser.text();
-                } else if ("_version".equals(currentFieldName) || "version".equals(currentFieldName)) {
-                    item.version = parser.longValue();
-                } else if ("_version_type".equals(currentFieldName) || "_versionType".equals(currentFieldName)
-                        || "version_type".equals(currentFieldName) || "versionType".equals(currentFieldName)) {
-                    item.versionType = VersionType.fromString(parser.text());
-                } else {
-                    throw new ElasticsearchParseException(
-                            "failed to parse More Like This item. unknown field [{}]", currentFieldName);
-                }
-            }
-        }
-        if (item.id != null && item.doc != null) {
-            throw new ElasticsearchParseException(
-                    "failed to parse More Like This item. either [id] or [doc] can be specified, but not both!");
-        }
-        return item;
-    }
-
-    @Override
-    public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
-        builder.startObject();
-        if (this.index != null) {
-            builder.field(Field.INDEX.getPreferredName(), this.index);
-        }
-        if (this.type != null) {
-            builder.field(Field.TYPE.getPreferredName(), this.type);
-        }
-        if (this.id != null && this.doc == null) {
-            builder.field(Field.ID.getPreferredName(), this.id);
-        }
-        if (this.doc != null) {
-            XContentType contentType = XContentFactory.xContentType(this.doc);
-            if (contentType == builder.contentType()) {
-                builder.rawField(Field.DOC.getPreferredName(), this.doc);
-            } else {
-                XContentParser parser = XContentFactory.xContent(contentType).createParser(this.doc);
-                parser.nextToken();
-                builder.field(Field.DOC.getPreferredName());
-                builder.copyCurrentStructure(parser);
-            }
-        }
-        if (this.fields != null) {
-            builder.array(Field.FIELDS.getPreferredName(), this.fields);
-        }
-        if (this.perFieldAnalyzer != null) {
-            builder.field(Field.PER_FIELD_ANALYZER.getPreferredName(), this.perFieldAnalyzer);
-        }
-        if (this.routing != null) {
-            builder.field(Field.ROUTING.getPreferredName(), this.routing);
-        }
-        if (this.version != Versions.MATCH_ANY) {
-            builder.field(Field.VERSION.getPreferredName(), this.version);
-        }
-        if (this.versionType != VersionType.INTERNAL) {
-            builder.field(Field.VERSION_TYPE.getPreferredName(), this.versionType.toString().toLowerCase(Locale.ROOT));
-        }
-        return builder.endObject();
-    }
-
-    @Override
-    public final String toString() {
-        try {
-            XContentBuilder builder = XContentFactory.jsonBuilder();
-            builder.prettyPrint();
-            toXContent(builder, EMPTY_PARAMS);
-            return builder.string();
-        } catch (Exception e) {
-            return "{ \"error\" : \"" + ExceptionsHelper.detailedMessage(e) + "\"}";
-        }
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(index, type, id, doc, Arrays.hashCode(fields), perFieldAnalyzer, routing,
-                version, versionType);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Item)) return false;
-        Item other = (Item) o;
-        return Objects.equals(index, other.index) &&
-                Objects.equals(type, other.type) &&
-                Objects.equals(id, other.id) &&
-                Objects.equals(doc, other.doc) &&
-                Arrays.equals(fields, other.fields) &&  // otherwise we are comparing pointers
-                Objects.equals(perFieldAnalyzer, other.perFieldAnalyzer) &&
-                Objects.equals(routing, other.routing) &&
-                Objects.equals(version, other.version) &&
-                Objects.equals(versionType, other.versionType);
-    }
-}
