+++ core/src/test/java/org/elasticsearch/search/morelikethis/MoreLikeThisIT.java
@@ -28,9 +28,9 @@ import org.elasticsearch.client.Client;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.xcontent.XContentBuilder;
 import org.elasticsearch.common.xcontent.XContentFactory;
+import org.elasticsearch.index.query.MoreLikeThisQueryBuilder;
+import org.elasticsearch.index.query.MoreLikeThisQueryBuilder.Item;
 import org.elasticsearch.index.query.QueryBuilders;
-import org.elasticsearch.index.query.morelikethis.Item;
-import org.elasticsearch.index.query.morelikethis.MoreLikeThisQueryBuilder;
 import org.elasticsearch.test.ESIntegTestCase;
 import org.junit.Test;
